Moved uniform_mesher to separate file
